Added hero spells endpoint

+++ src/main/java/gg/bayes/dota_challenge/adapter/dto/HeroSpells.java
@@ -1,5 +1,5 @@
 package gg.bayes.dota_challenge.adapter.dto;
 
 // Using records for DTO classes (Records are available since JDK 14)
-public record HeroSpells(String spell, int casts) {
+public record HeroSpells(String spell, long casts) {
 }
